[block] Move func (b *Block) ActionHashs() into block_test.go
